Tutorial: DICOM to USD with MONAI Deploy and Holoscan
Adds a tutorial demonstrating how to ingest DICOM data, run inference
with MONAI Deploy, convert to STL, and then write to disk as USD with
Holoscan.

This tutorial is adapted from prior work to demonstrate a workflow
from MONAI Deploy to the NVIDIA Omniverse cloud platform. OpenUSD
presets a bridge format between edge-based applications with Holoscan and MONAI
Deploy and cloud-based applications in NVIDIA Omniverse, though the
tutorial focuses solely on a local workflow for now.

The tutorial also demonstrates seamless integration between Holoscan and
MONAI Deploy in a single application.

Depends on the Holoscan 1.0.3 container. Runs on amd64 and arm64
platforms.

co-authored-by: Cristiana Dinea <cdinea@nvidia.com>
co-authored-by: Wendell Hom <whom@nvidia.com>
Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>