Utility: Check in script to gather JSON metadata
Add `gather_metadata.py` to Git tracking on the `main` branch to avoid
duplication.

`gather_metadata.py` was previously introduced and tracked on the
`github.io` branch. It is added to the `main` branch here to avoid
duplication of utility efforts as we investigate paths to increase
visibility of HoloHub projects.

co-authored-by: Julien Jomier <jjomier@nvidia.com>
Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>