Utility: Add missing schema for "run" command
Adds missing JSON schema for the command to run an application.

The "run" command metadata field has been implicitly part of the
application schema for some time and is referenced in the `run` script
along with most applications, but was missing from JSON schema
validation prior to this change.

The "run" command field is optional. Apps with complex run instructions
may choose not to provide a "run" command and instead provide manual
"run" steps in their README.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>