Utility: Add optional tutorial schema (v1)
Adds a JSON schema for HoloHub tutorial metadata based on the existing
HoloHub application schema.

To reflect the fact that tutorials prior to this change have not been
required to provide metadata nor had a standard to follow for doing so,
the metadata validation utility has been updated to warn instead of
error in the case that a tutorial is missing metadata.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>