Tutorial: Add flow benchmarking support for Python applications
Adds support for benchmarking Python applications in HoloHub as a
parameter to the Holoscan Flow Benchmarking tutorial `benchmark.py`
script.

Under previous behavior the benchmarking script primarily supported only
HoloHub C++ applications. The recommended path to benchmark Python
applications was to pass the full command string to "--run-command". This
approach disallowed use of the "run" script that provides infrastructure
for environment setup and working directory management. As a result, in
practice "--run-command" alone required manual setup of the environment
and working directory, i.e. several extra steps on the user's part.

This update allows users to launch a Python application with the "run"
script to minimize additional user effort. The "--language" argument is
forwarded to the "run" script which executes as otherwise recommended in
HoloHub instructions. The "--run-command" parameter is left intact.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>